TST: GH18498 Requested pull request changes
  Consolidate functions to reduce amount of module files.
